+++ src/contents/settings/file-formats.tsx
@@ -5,8 +5,9 @@ import { useSettings } from "@/context/settings";
 import { NEW_TRANSLATION_FORMATS } from "@/lib/constants/items";
 import { TranslationFormat } from "@/lib/types/enums";
 import { Files } from "lucide-react";
+import { Switch } from "@/components/ui/switch";
 import { lazy, Suspense } from "react";
-import { SiAndroidstudio } from "react-icons/si";
+import { SiAndroid } from "react-icons/si";
 
 const SelectorField = lazy(()=>import("@/components/fields/selector"))
 
@@ -35,15 +36,57 @@ export default function FileFormatSettings(){
                </SettingsItem>
                <SettingsItem
                     title="Android XML"
-                    Icon={SiAndroidstudio}
+                    Icon={SiAndroid}
                >
-                    TODO: Default translatable state, Show translatable column
+                    <SettingsOption
+                         title="Default Translatable State"
+                         description="Initial value of the Android translatable state"
+                         id="default-translatable"
+                    >
+                         <Switch
+                              id="default-translatable"
+                              checked={settings.androidTranslatable}
+                              onCheckedChange={checked=>setSettings({ androidTranslatable: checked })}
+                         />
+                    </SettingsOption>
+                    <SettingsOption
+                         title="Show Translatable column"
+                         description="Show translatable table column every time you enter the app"
+                         id="show-translatable"
+                    >
+                         <Switch
+                              id="show-translatable"
+                              checked={settings.androidTranslatableCol}
+                              onCheckedChange={checked=>setSettings({ androidTranslatableCol: checked })}
+                         />
+                    </SettingsOption>
                </SettingsItem>
                <SettingsItem
                     title="XLIFF"
                     Icon={Files}
                >
-                    TODO: Auto-detect languages, Preserve metadata
+                    <SettingsOption
+                         title="Auto-detect language"
+                         description="Auto-detect language everytime the XLIFF file is opened"
+                         id="auto-detect-lang"
+                    >
+                         <Switch
+                              id="auto-detect-lang"
+                              checked={settings.xliffAutoDetect}
+                              onCheckedChange={checked=>setSettings({xliffAutoDetect: checked})}
+                         />
+                    </SettingsOption>
+                    <SettingsOption
+                         title="Preserve Metadata"
+                         description="Keep existing translation metadata when opening or saving XLIFF files"
+                         id="preserve-meta"
+                    >
+                         <Switch
+                              id="preserve-meta"
+                              checked={settings.xliffPreserveMeta}
+                              onCheckedChange={checked => setSettings({ xliffPreserveMeta: checked })}
+                         />
+                    </SettingsOption>
                </SettingsItem>
           </div>
      )