+++ src/contents/settings/shortcuts.tsx
@@ -6,7 +6,7 @@ const ShortcutsTable = lazy(()=>import("@/components/tables/shortcuts"))
 export default function ShortcutSettings(){
      return (
           <div className="space-y-2">
-               <Suspense fallback={<TableLoader rows={10} cols={[50,100]}/>}>
+               <Suspense fallback={<TableLoader rows={10} cols={[50,100]} excludeMinWidth/>}>
                     <ShortcutsTable/>
                </Suspense>
           </div>