+++ src/contents/settings/translation.tsx
@@ -18,7 +18,7 @@ export default function TranslationSettings(){
                          id="gemini-key"
                     >
                          <Input
-                              name="gemini-key"
+                              id="gemini-key"
                               className="w-1/3"
                               value={providers.geminiApi}
                               onChange={e=>setProviders({geminiApi: e.target.value.trim()})}
@@ -35,7 +35,7 @@ export default function TranslationSettings(){
                          id="libre-translate-server"
                     >
                          <Input
-                              name="libre-translate-server"
+                              id="libre-translate-server"
                               className="w-1/3"
                               value={providers.libreTranslateServer}
                               onChange={e=>setProviders({libreTranslateServer: e.target.value.trim()})}
@@ -47,7 +47,7 @@ export default function TranslationSettings(){
                          id="libre-translate-key"
                     >
                          <Input
-                              name="libre-translate-key"
+                              id="libre-translate-key"
                               className="w-1/3"
                               value={providers.libreTranslateApi}
                               onChange={e=>setProviders({libreTranslateApi: e.target.value.trim()})}
@@ -64,7 +64,7 @@ export default function TranslationSettings(){
                          id="llama-endpoint"
                     >
                          <Input
-                              name="llama-endpoint"
+                              id="llama-endpoint"
                               className="w-1/3"
                               value={providers.llamaEndpoint}
                               onChange={e=>setProviders({llamaEndpoint: e.target.value.trim()})}
@@ -76,14 +76,13 @@ export default function TranslationSettings(){
                          id="llama-model"
                     >
                          <Input
-                              name="llama-model"
+                              id="llama-model"
                               className="w-1/3"
                               value={providers.llamaModel}
                               onChange={e=>setProviders({llamaModel: e.target.value})}
                          />
                     </SettingsOption>
                </SettingsItem>
-               {/* TODO: Preserve empty translations */}
           </div>
      )
 }