+++ src/loaders/settings/shortcuts.tsx
@@ -3,7 +3,7 @@ import { TableLoader } from "../translator";
 export default function ShortcutSettings(){
      return (
           <div className="space-y-2">
-               <TableLoader rows={10} cols={[50,100]}/>
+               <TableLoader rows={10} cols={[50,100]} excludeMinWidth/>
           </div>
      )
 }